edit durationOnBoardingScreen

+++ lib/config/constants.dart
@@ -2,7 +2,7 @@ class Constants {
   static const double designSizeWidth = 390;
   static const double designSizeHeight = 844;
   static const double elevationAppBar = 0;
-  static const double durationOnBoardingScreen = 1;
+  static const int durationOnBoardingScreen = 1;
   static const double blurRadiusOfBoxShadowToBottomContainerInOnBoardingScreen =
       2.73;
   static const double
